Fix theme admin test
