--- bluebottle/bb_projects/tests/test_admin.py
@@ -1,24 +1,0 @@
-# -*- coding: utf-8 -*-
-from django.contrib.admin.sites import AdminSite
-from django.urls.base import reverse
-from rest_framework.status import HTTP_200_OK
-
-from bluebottle.initiatives.admin import ThemeAdmin
-from bluebottle.initiatives.tests.factories import InitiativeFactory
-from bluebottle.time_based.models import Skill
-from bluebottle.test.utils import BluebottleAdminTestCase
-
-
-class TestThemeAdmin(BluebottleAdminTestCase):
-
-    def setUp(self):
-        super(TestThemeAdmin, self).setUp()
-        self.site = AdminSite()
-        self.skill_admin = ThemeAdmin(Skill, self.site)
-        self.client.force_login(self.superuser)
-        InitiativeFactory.create()
-
-    def test_theme_admin_list(self):
-        url = reverse('admin:bb_projects_projecttheme_changelist')
-        response = self.client.get(url)
-        self.assertTrue(response, HTTP_200_OK)
